+++ src/apmodel/vocab/activity.py
@@ -1,4 +1,6 @@
-from ..core import Activity
+from ..core import Activity, Object, Link
+
+from ..funcs import _make_accept, _make_reject
 
 class Accept(Activity):
     def __init__(self, **kwargs):
@@ -6,13 +8,11 @@ class Accept(Activity):
             kwargs.pop("type")
         super().__init__("Accept", **kwargs)
 
-
 class Reject(Activity):
     def __init__(self, **kwargs):
         if kwargs.get("type"):
             kwargs.pop("type")
         super().__init__("Reject", **kwargs)
-
 
 class TentativeReject(Activity):
     def __init__(self, **kwargs):
@@ -20,6 +20,13 @@ class TentativeReject(Activity):
             kwargs.pop("type")
         super().__init__("TentativeReject", **kwargs)
 
+    def accept(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_accept(obj, actor)
+
+    def reject(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_reject(obj, actor)
 
 class Remove(Activity):
     def __init__(self, **kwargs):
@@ -27,6 +34,13 @@ class Remove(Activity):
             kwargs.pop("type")
         super().__init__("Remove", **kwargs)
 
+    def accept(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_accept(obj, actor)
+
+    def reject(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_reject(obj, actor)
 
 class Undo(Activity):
     def __init__(self, **kwargs):
@@ -34,6 +48,13 @@ class Undo(Activity):
             kwargs.pop("type")
         super().__init__("Undo", **kwargs)
 
+    def accept(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_accept(obj, actor)
+
+    def reject(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_reject(obj, actor)
 
 class Create(Activity):
     def __init__(self, **kwargs):
@@ -41,6 +62,13 @@ class Create(Activity):
             kwargs.pop("type")
         super().__init__("Create", **kwargs)
 
+    def accept(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_accept(obj, actor)
+
+    def reject(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_reject(obj, actor)
 
 class Delete(Activity):
     def __init__(self, **kwargs):
@@ -48,6 +76,13 @@ class Delete(Activity):
             kwargs.pop("type")
         super().__init__("Delete", **kwargs)
 
+    def accept(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_accept(obj, actor)
+
+    def reject(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_reject(obj, actor)
 
 class Update(Activity):
     def __init__(self, **kwargs):
@@ -55,6 +90,13 @@ class Update(Activity):
             kwargs.pop("type")
         super().__init__("Update", **kwargs)
 
+    def accept(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_accept(obj, actor)
+
+    def reject(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_reject(obj, actor)
 
 class Follow(Activity):
     def __init__(self, **kwargs):
@@ -62,6 +104,13 @@ class Follow(Activity):
             kwargs.pop("type")
         super().__init__("Follow", **kwargs)
 
+    def accept(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_accept(obj, actor)
+
+    def reject(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_reject(obj, actor)
 
 class View(Activity):
     def __init__(self, **kwargs):
@@ -69,6 +118,13 @@ class View(Activity):
             kwargs.pop("type")
         super().__init__("View", **kwargs)
 
+    def accept(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_accept(obj, actor)
+
+    def reject(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_reject(obj, actor)
 
 class Listen(Activity):
     def __init__(self, **kwargs):
@@ -83,6 +139,13 @@ class Read(Activity):
             kwargs.pop("type")
         super().__init__("Read", **kwargs)
 
+    def accept(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_accept(obj, actor)
+
+    def reject(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_reject(obj, actor)
 
 class Move(Activity):
     def __init__(self, **kwargs):
@@ -90,6 +153,13 @@ class Move(Activity):
             kwargs.pop("type")
         super().__init__("Move", **kwargs)
 
+    def accept(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_accept(obj, actor)
+
+    def reject(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_reject(obj, actor)
 
 class Travel(Activity):
     def __init__(self, **kwargs):
@@ -97,6 +167,13 @@ class Travel(Activity):
             kwargs.pop("type")
         super().__init__("Travel", **kwargs)
 
+    def accept(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_accept(obj, actor)
+
+    def reject(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_reject(obj, actor)
 
 class Announce(Activity):
     def __init__(self, **kwargs):
@@ -104,6 +181,13 @@ class Announce(Activity):
             kwargs.pop("type")
         super().__init__("Announce", **kwargs)
 
+    def accept(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_accept(obj, actor)
+
+    def reject(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_reject(obj, actor)
 
 class Block(Activity):
     def __init__(self, **kwargs):
@@ -111,6 +195,13 @@ class Block(Activity):
             kwargs.pop("type")
         super().__init__("Block", **kwargs)
 
+    def accept(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_accept(obj, actor)
+
+    def reject(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_reject(obj, actor)
 
 class Flag(Activity):
     def __init__(self, **kwargs):
@@ -118,18 +209,42 @@ class Flag(Activity):
             kwargs.pop("type")
         super().__init__("Flag", **kwargs)
 
+    def accept(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_accept(obj, actor)
+
+    def reject(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_reject(obj, actor)
+    
 class Like(Activity):
     def __init__(self, **kwargs):
         if kwargs.get("type"):
             kwargs.pop("type")
         super().__init__("Like", **kwargs)
 
+    def accept(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_accept(obj, actor)
+
+    def reject(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_reject(obj, actor)
+    
 class Dislike(Activity):
     def __init__(self, **kwargs):
         if kwargs.get("type"):
             kwargs.pop("type")
         super().__init__("Dislike", **kwargs)
 
+    def accept(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_accept(obj, actor)
+
+    def reject(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_reject(obj, actor)
+    
 class IntransitiveActivity(Activity):
     def __init__(
         self,
@@ -145,6 +260,13 @@ class IntransitiveActivity(Activity):
             "IntransitiveActivity" if not type else type, actor, target, result, origin, instrument, **kwargs
         )
 
+    def accept(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_accept(obj, actor)
+
+    def reject(self, actor: Object | Link | str):
+        obj = self.to_dict(self._extras)
+        return _make_reject(obj, actor)
 
 class Question(IntransitiveActivity):
     def __init__(self, **kwargs):