+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.2.1
+## What's new
+- feat: `Activity.accept()` can now generate an Accept activity for the activity and `Activity.reject()` can generate a Reject activity for the activity
+# 0.2.1
+## What's new
+- feat: add `Multikey` support in actor
 # 0.2.0
 ## What's new
 - feat: add `Multikey` support in actor